tests: Remove unnecessary cast
The value `i` is already a `u8`. Addresses a clippy warning.

+++ star/tests/e2e.rs
@@ -268,7 +268,7 @@ fn star_with_aux_multiple_block(oprf_server: Option<PPOPRFServer>) {
     } else {
       // Unique measurements which will not meet threshold.
       clients.push(MessageGenerator::new(
-        SingleMeasurement::new(&[i as u8]),
+        SingleMeasurement::new(&[i]),
         threshold,
         epoch.as_bytes(),
       ));
